Rewrote process of the script
Rewrote the process of the script to have it download necessary files on
demand rather than having it create them itself. Reduced redundant code
pieces and tried to use less commands for the same amount of work.
Added a remove functionality that removes BitTorrent Sync from the
Raspberry Pi. I also added a new feature to remove the installation